Compacted the test app into one file
Well, really 3. There's also the one view file and the `config.ru`
