+++ config.ru
@@ -1,0 +1,6 @@
+# frozen_string_literal: true
+
+require_relative 'test_app'
+
+run Rails.application
+Rails.application.load_server
