--- test-app/Rakefile
@@ -1,4 +1,0 @@
-# frozen_string_literal: true
-
-require File.expand_path('config/application', __dir__)
-Rails.application.load_tasks
