--- test-app/bin/bundle
@@ -1,3 +1,0 @@
-#!/usr/bin/env ruby
-ENV['BUNDLE_GEMFILE'] ||= File.expand_path('..', __dir__)
-load Gem.bin_path('bundler', 'bundle')
