--- test-app/bin/rake
@@ -1,4 +1,0 @@
-#!/usr/bin/env ruby
-require_relative '../config/boot'
-require 'rake'
-Rake.application.run
