--- test-app/config.ru
@@ -1,3 +1,0 @@
-require_relative 'config/environment'
-
-run Rails.application
