--- test-app/config/environment.rb
@@ -1,2 +1,0 @@
-require_relative 'application'
-Rails.application.initialize!
