--- test-app/config/initializers/prawn_rails.rb
@@ -1,4 +1,0 @@
-PrawnRails.config do |c|
-  c.page_layout = :landscape
-  c.margin = 30
-end
