--- test-app/config/routes.rb
@@ -1,3 +1,0 @@
-Rails.application.routes.draw do
-  root 'pdf#static.pdf'
-end
