--- test-app/config/secrets.yml
@@ -1,2 +1,0 @@
-development:
-  secret_key_base: 6f8f8076169dafe91f4f8a61de1dc5adeaa4467454b1220cc2fe7170bea58c604e6274dcfd7abd634cf7c0fba5d0005736ab6f23e582f279e9cbf700867047b3
